refactor: s3 관련 로직을 common 에서 infra로 이동
