--- src/main/java/one/colla/common/application/dto/request/DomainType.java
@@ -1,6 +1,0 @@
-package one.colla.common.application.dto.request;
-
-public enum DomainType {
-	USER,
-	TEAMSPACE
-}
