+++ src/main/java/one/colla/infra/s3/application/FileService.java
@@ -1,4 +1,4 @@
-package one.colla.common.application;
+package one.colla.infra.s3.application;
 
 import java.net.URL;
 import java.util.Date;
@@ -15,13 +15,13 @@ import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest;
 
 import lombok.RequiredArgsConstructor;
-import one.colla.common.application.dto.request.DomainType;
-import one.colla.common.application.dto.request.FileUploadDto;
-import one.colla.common.application.dto.request.PreSignedUrlRequest;
-import one.colla.common.application.dto.response.FileUploadUrlsDto;
-import one.colla.common.application.dto.response.PreSignedUrlResponse;
 import one.colla.common.security.authentication.CustomUserDetails;
-import one.colla.common.util.S3Util;
+import one.colla.infra.s3.application.dto.request.DomainType;
+import one.colla.infra.s3.application.dto.request.FileUploadDto;
+import one.colla.infra.s3.application.dto.request.PreSignedUrlRequest;
+import one.colla.infra.s3.application.dto.response.FileUploadUrlsDto;
+import one.colla.infra.s3.application.dto.response.PreSignedUrlResponse;
+import one.colla.infra.s3.util.S3Util;
 import one.colla.teamspace.application.TeamspaceService;
 
 @Service
