+++ src/main/java/one/colla/infra/s3/application/dto/request/DomainType.java
@@ -1,0 +1,6 @@
+package one.colla.infra.s3.application.dto.request;
+
+public enum DomainType {
+	USER,
+	TEAMSPACE
+}
