+++ src/main/java/one/colla/infra/s3/presentation/S3Controller.java
@@ -1,4 +1,4 @@
-package one.colla.common.presentation;
+package one.colla.infra.s3.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -10,10 +10,11 @@ import org.springframework.web.bind.annotation.RestController;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import one.colla.common.application.FileService;
-import one.colla.common.application.dto.request.PreSignedUrlRequest;
-import one.colla.common.application.dto.response.PreSignedUrlResponse;
+import one.colla.common.presentation.ApiResponse;
 import one.colla.common.security.authentication.CustomUserDetails;
+import one.colla.infra.s3.application.FileService;
+import one.colla.infra.s3.application.dto.request.PreSignedUrlRequest;
+import one.colla.infra.s3.application.dto.response.PreSignedUrlResponse;
 
 @RestController
 @RequestMapping("/api/v1")
